services/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.32
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openai==1.50.0
httpx
python-multipart==0.0.9
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
pytest
